ui: show btm sheet fully on foldable/tablets
