+++ app/src/main/java/com/celzero/bravedns/ui/bottomsheet/RethinkLogBottomSheet.kt
@@ -48,7 +48,6 @@ import com.celzero.bravedns.util.UIUtils.htmlToSpannedText
 import com.celzero.bravedns.util.Utilities
 import com.celzero.bravedns.util.Utilities.getIcon
 import com.celzero.bravedns.util.Utilities.showToastUiCentered
-import com.celzero.bravedns.util.useTransparentNoDimBackground
 import com.google.gson.Gson
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -78,11 +77,6 @@ class RethinkLogBottomSheet : BaseBottomSheetDialogFragment(), KoinComponent {
     ): View {
         _binding = BottomSheetConnTrackBinding.inflate(inflater, container, false)
         return b.root
-    }
-
-    override fun onStart() {
-        super.onStart()
-        dialog?.useTransparentNoDimBackground()
     }
 
     override fun onDestroyView() {
